+++ maze_game_final/Assets/Prefabs/Collectables/Coffee.prefab
@@ -167,7 +167,7 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   collectablesContainer: {fileID: 11400000, guid: fc4644a0f31f43b4d8031d98c2f4f47d, type: 2}
   delay: 3
-  speed: 10
+  speed: 25
   playerSpeed: {fileID: 11400000, guid: 8d817ce0900df834a89c82316037e9e7, type: 2}
 --- !u!114 &1379524086889285923
 MonoBehaviour:
