+++ maze_game_final/Assets/Prefabs/Corrounding/Boss Light Pref.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 20cc0faf38e3c8646b0a966c0f0de259
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
