--- maze_game_final/Assets/Scripts/Level.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: ae6fbbd5a3719f24fb0ba18c52737fe7
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
