--- maze_game_final/Assets/Scripts/Level/LevelController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 664780ac0436e15469c8498bbb69f644
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
