--- maze_game_final/Assets/Scripts/Level/LevelView.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f5afd586a02ac2b48a057080bca1da97
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
