--- maze_game_final/Assets/Scripts/Level/Wall.cs
@@ -1,9 +1,0 @@
-namespace Level
-{
-    using UnityEngine;
-
-    public class Wall : MonoBehaviour
-    {
-        
-    }
-}
