+++ maze_game_final/Assets/Scripts/Maze/LevelCell.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 869238fa2a677914d89a4cd1a9c51e70
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
