+++ maze_game_final/Assets/Scripts/Maze/Wall.cs
@@ -1,0 +1,9 @@
+namespace Level
+{
+    using UnityEngine;
+
+    public class Wall : MonoBehaviour
+    {
+        
+    }
+}
