+++ maze_game_final/Assets/Scripts/Maze/Wall.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1d4ebb96c20093742be88cd370b1e7aa
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
