+++ maze_game_final/Assets/Scripts/Maze/WallSO.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 29ad890ebbb8d9d44876b00cc30a425a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
