+++ maze_game_final/Assets/Scripts/Maze/WallView.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5f57650fc57753b45bfe93bba2ef807a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
