+++ maze_game_final/Assets/Scripts/Surroundings/BossLights.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9a02e74477f7a9a4aaf04cf38d74b7f0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
