+++ maze_game_final/Assets/Scripts/Timers/CustomTimer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 04416550d1703be429b94089b63bc0a4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
